Fixed changes for the review
