+++ lib/board.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class Board
-  attr_accessor :board
+  attr_reader :board
   attr_reader :turn
   attr_reader :player_one_turn
 
@@ -11,8 +11,18 @@ class Board
     @player_one_turn = true
   end
 
-  def check_for_winner(boardstate)
-    current_b = boardstate[0] + boardstate[1] + boardstate[2]
+  def selection(player, num)
+    @board.each do |n|
+      next unless n.include? num
+
+      n.map! do |ele|
+        ele == num ? player : ele
+      end
+    end
+  end
+
+  def check_for_winner(_board_elem)
+    current_b = @board[0] + @board[1] + @board[2]
 
     return [true, current_b[0]] if current_b[0] == current_b[1] && current_b[1] == current_b[2]
     return [true, current_b[3]] if current_b[3] == current_b[4] && current_b[4] == current_b[5]
@@ -37,5 +47,39 @@ class Board
 
   def switch_player
     @player_one_turn = !@player_one_turn
+  end
+
+  def valid?(num)
+    selec_index = convert_to_index(num)
+    col = []
+    @board.each_with_index do |n, index|
+      col = n if index == selec_index[0]
+    end
+
+    if col.include_any?(%w[X O])
+
+      col.each_with_index do |x, i|
+        return false if x == 'X' && i == selec_index[1] || x == 'O' && i == selec_index[1]
+      end
+    end
+    true
+  end
+
+  def convert_to_index(num)
+    return [0, 0] if num == '1'
+    return [0, 1] if num == '2'
+    return [0, 2] if num == '3'
+    return [1, 0] if num == '4'
+    return [1, 1] if num == '5'
+    return [1, 2] if num == '6'
+    return [2, 0] if num == '7'
+    return [2, 1] if num == '8'
+    return [2, 2] if num == '9'
+  end
+end
+
+module Enumerable
+  def include_any?(array)
+    array.any? { |i| include? i }
   end
 end
